fix: resolve clippy warnings and egui 0.29 API mismatches
- Replace Color32::lighten (non-existent) with gamma_multiply
- Replace Shape::triangle/convex_polygon struct with Shape::convex_polygon fn
- Replace Slider::desired_width (non-existent) — use defaults
- Replace RadioButton::fill (non-existent) with separate color swatch
- Fix f32/f64 type mismatches in timeline drag math
- Fix f32 type for ProgressBar::new
- Fix HashMap<f64, TextureHandle> — f64 isn't Hash; use u64 frame index
- Fix borrow-checker issue in menubar Playback menu
- Fix Memory::has_focus (needs Id arg) — use focused().is_some()
- Fix tuple destructuring in preview panel
- Fix Field-reassign-with-default in Project::new
- Mark run_ffmpeg as #[allow(dead_code)] (used by future code paths)
- All 51 unit tests pass
- cargo fmt --all applied
- cargo clippy --workspace --all-targets clean

+++ crates/edt-app/src/ui/export_dialog.rs
@@ -260,7 +260,7 @@ fn render_progress(app: &mut EdtApp, ui: &mut Ui) {
     let frac = if prog.frames_total == 0 {
         0.0
     } else {
-        prog.fraction()
+        prog.fraction() as f32
     };
     ui.add(egui::ProgressBar::new(frac).text(format!(
         "{}%  ({}/{})",
